+++ lib/THNN/generic/SpatialConvolutionLocal.c
@@ -5,8 +5,8 @@
 static inline void THNN_(SpatialConvolutionLocal_shapeCheck)(
 	THTensor *input, THTensor *gradOutput,
 	THTensor *weight, THTensor *bias, 
-	int kH, int kW, int dH, 
-	int dW, int padH, int padW,
+	intptr_t kH, intptr_t kW, intptr_t dH,
+	intptr_t dW, intptr_t padH, intptr_t padW,
 	long inputHeight, long inputWidth,
 	long outputHeight, long outputWidth) {
 
@@ -110,9 +110,9 @@ void THNN_(SpatialConvolutionLocal_updateOutput)(
     THTensor *bias,
     THTensor *finput,
     THTensor *fgradInput,
-    int kW, int kH,
-    int dW, int dH,
-    int padW, int padH,
+    intptr_t kW, intptr_t kH,
+    intptr_t dW, intptr_t dH,
+    intptr_t padW, intptr_t padH,
     long inputWidth, long inputHeight,
     long outputWidth, long outputHeight)
 {
@@ -211,9 +211,9 @@ void THNN_(SpatialConvolutionLocal_updateGradInput)(
     THTensor *weight,
     THTensor *finput,
     THTensor *fgradInput,
-    int kW, int kH,
-    int dW, int dH,
-    int padW, int padH,
+    intptr_t kW, intptr_t kH,
+    intptr_t dW, intptr_t dH,
+    intptr_t padW, intptr_t padH,
     long inputWidth, long inputHeight,
     long outputWidth, long outputHeight)
 {
@@ -274,7 +274,7 @@ void THNN_(SpatialConvolutionLocal_updateGradInput)(
 static void THNN_(SpatialConvolutionLocal_accGradParameters_frame)
      (THTensor *gradOutput, THTensor *gradWeight, THTensor *gradBias,
       THTensor *finput, real scale, 
-      int kW, int kH, int dW, int dH, int padW, int padH, 
+      intptr_t kW, intptr_t kH, intptr_t dW, intptr_t dH, intptr_t padW, intptr_t padH,
       long nInputPlane, long inputWidth, long inputHeight,
       long nOutputPlane, long outputWidth, long outputHeight)
 {
@@ -307,9 +307,9 @@ void THNN_(SpatialConvolutionLocal_accGradParameters)(
     THTensor *gradBias,
     THTensor *finput,
     THTensor *fgradInput,
-    int kW, int kH,
-    int dW, int dH,
-    int padW, int padH,
+    intptr_t kW, intptr_t kH,
+    intptr_t dW, intptr_t dH,
+    intptr_t padW, intptr_t padH,
     long inputWidth, long inputHeight,
     long outputWidth, long outputHeight,
     real scale)
